textdetection/requirements.txt
streamlit==1.26.0
python-decouple==3.8
Pillow==9.4.0
annotated-text==
openai==0.28.0
python-dotenv==1.0.0